backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-multipart
python-dotenv
google-genai
pymupdf
pillow
psycopg2-binary
python-docx
requests
